API

DACE.DA

DACE.DA

DACE.DA

DACE.deriv

DACE.deriv

DACE.eval

DACE.eval

DACE.eval

DACE.eval

DACE.eval

DACE.eval

DACE.eval

DACE.eval

DACE.eval

DACE.eval

DACE.eval

DACE.eval

DACE.eval

DACE.eval

DACE.eval

DACE.eval

DACE.evalScalar

DACE.evalScalar

DACE.evalScalar

DACE.evalScalar

DACE.evalScalar

DACE.evalScalar

DACE.getCentralMoments

DACE.getCoefficient

DACE.getCoefficient

DACE.getCoefficient

DACE.getCoefficient

DACE.getEps

DACE.getEpsMac

DACE.getExponents

DACE.getMaxMonomials

DACE.getMaxOrder

DACE.getMaxVariables

DACE.getMonomial

DACE.getMonomials

DACE.getMultiIndices

DACE.getRawMoments

DACE.getTO

DACE.hessian

DACE.identity

DACE.identity

DACE.identity

DACE.init

DACE.popTO

DACE.pushTO

DACE.setCoefficient!

DACE.setCoefficient!

DACE.setCoefficient!

DACE.setEps

DACE.setTO

DACE.DA — Type

abstract type DA <: Real

DA object representing a single polynomial.

The default constructor (with no arguments) creates an empty DA object representing the constant zero function.

DACE.DA — Method

DA(i::Integer, c::Number)

Create a DA object as c times the independent variable number i.

DACE.DA — Method

DA(c::Number)

Create a DA object with the constant part equal to c.

DACE.deriv — Method

deriv(
    arg1::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}},
    arg2::Integer
) -> DACE.DAAllocated

Derivative of DA arg1 with respect to given variable, arg2

DACE.deriv — Method

deriv(
    arg1::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}},
    arg2::Union{CxxWrap.CxxWrapCore.ConstCxxRef{CxxWrap.StdLib.StdVector{UInt32}}, CxxWrap.CxxWrapCore.CxxRef{CxxWrap.StdLib.StdVector{UInt32}}, CxxWrap.StdLib.StdVector{UInt32}}
) -> DACE.DAAllocated

Derivative of DA arg1 with respect to given variables, arg2

DACE.eval — Method

eval(
    arg1::Union{AlgebraicVector{DA}, CxxWrap.CxxWrapCore.ConstCxxRef{<:AlgebraicVector{DA}}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{DA}}},
    arg2::Union{AlgebraicVector{DA}, CxxWrap.CxxWrapCore.ConstCxxRef{<:AlgebraicVector{DA}}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{DA}}}
) -> DACE.AlgebraicVectorAllocated{DA}

Evaluation of arg1 with a vector of arguments, arg2.

DACE.eval — Method

eval(
    arg1::Union{AlgebraicVector{DA}, CxxWrap.CxxWrapCore.ConstCxxRef{<:AlgebraicVector{DA}}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{DA}}},
    arg2::Union{AlgebraicVector{Float64}, CxxWrap.CxxWrapCore.ConstCxxRef{<:AlgebraicVector{Float64}}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{Float64}}}
) -> DACE.AlgebraicVectorAllocated{Float64}

Evaluation of arg1 with a vector of arguments, arg2.

DACE.eval — Method

eval(
    arg1::Union{AlgebraicVector{DA}, CxxWrap.CxxWrapCore.ConstCxxRef{<:AlgebraicVector{DA}}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{DA}}},
    arg2::Union{CxxWrap.StdLib.StdVector{DA}, CxxWrap.CxxWrapCore.ConstCxxRef{<:CxxWrap.StdLib.StdVector{DA}}, CxxWrap.CxxWrapCore.CxxRef{<:CxxWrap.StdLib.StdVector{DA}}}
) -> CxxWrap.StdLib.StdVectorAllocated{DA}

Evaluation of arg1 with a vector of arguments, arg2.

DACE.eval — Method

eval(
    arg1::Union{AlgebraicVector{DA}, CxxWrap.CxxWrapCore.ConstCxxRef{<:AlgebraicVector{DA}}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{DA}}},
    arg2::Union{CxxWrap.StdLib.StdVector{Float64}, CxxWrap.CxxWrapCore.ConstCxxRef{<:CxxWrap.StdLib.StdVector{Float64}}, CxxWrap.CxxWrapCore.CxxRef{<:CxxWrap.StdLib.StdVector{Float64}}}
) -> CxxWrap.StdLib.StdVectorAllocated{Float64}

Evaluation of arg1 with a vector of arguments, arg2.

DACE.eval — Method

eval(
    arg1::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}},
    arg2::Union{AlgebraicVector{DA}, CxxWrap.CxxWrapCore.ConstCxxRef{<:AlgebraicVector{DA}}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{DA}}}
) -> DACE.DAAllocated

Evaluation of arg1 with a vector of arguments, arg2.

DACE.eval — Method

eval(
    arg1::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}},
    arg2::Union{AlgebraicVector{Float64}, CxxWrap.CxxWrapCore.ConstCxxRef{<:AlgebraicVector{Float64}}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{Float64}}}
) -> Float64

Evaluation of arg1 with a vector of arguments, arg2.

DACE.eval — Method

eval(
    arg1::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}},
    arg2::Union{CxxWrap.StdLib.StdVector{DA}, CxxWrap.CxxWrapCore.ConstCxxRef{<:CxxWrap.StdLib.StdVector{DA}}, CxxWrap.CxxWrapCore.CxxRef{<:CxxWrap.StdLib.StdVector{DA}}}
) -> DACE.DAAllocated

Evaluation of arg1 with a vector of arguments, arg2.

DACE.eval — Method

eval(
    arg1::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}},
    arg2::Union{CxxWrap.StdLib.StdVector{Float64}, CxxWrap.CxxWrapCore.ConstCxxRef{<:CxxWrap.StdLib.StdVector{Float64}}, CxxWrap.CxxWrapCore.CxxRef{<:CxxWrap.StdLib.StdVector{Float64}}}
) -> Float64

Evaluation of arg1 with a vector of arguments, arg2.

DACE.eval — Method

eval(
    arg1::Union{compiledDA, CxxWrap.CxxWrapCore.ConstCxxRef{<:compiledDA}, CxxWrap.CxxWrapCore.CxxRef{<:compiledDA}},
    arg2::Union{AlgebraicVector{DA}, CxxWrap.CxxWrapCore.ConstCxxRef{<:AlgebraicVector{DA}}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{DA}}},
    arg3::Union{AlgebraicVector{DA}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{DA}}}
)

Evaluate the compiled polynomial, arg1, with a vector of any arithmetic type, arg2, and return vector of results, arg3.

DACE.eval — Method

eval(
    arg1::Union{compiledDA, CxxWrap.CxxWrapCore.ConstCxxRef{<:compiledDA}, CxxWrap.CxxWrapCore.CxxRef{<:compiledDA}},
    arg2::Union{AlgebraicVector{DA}, CxxWrap.CxxWrapCore.ConstCxxRef{<:AlgebraicVector{DA}}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{DA}}}
)

Evaluation of arg1 with a vector of arguments, arg2.

DACE.eval — Method

eval(
    arg1::Union{compiledDA, CxxWrap.CxxWrapCore.ConstCxxRef{<:compiledDA}, CxxWrap.CxxWrapCore.CxxRef{<:compiledDA}},
    arg2::Union{AlgebraicVector{Float64}, CxxWrap.CxxWrapCore.ConstCxxRef{<:AlgebraicVector{Float64}}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{Float64}}},
    arg3::Union{AlgebraicVector{Float64}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{Float64}}}
)

Evaluate the compiled polynomial, arg1, with a vector of any arithmetic type, arg2, and return vector of results, arg3.

DACE.eval — Method

eval(
    arg1::Union{compiledDA, CxxWrap.CxxWrapCore.ConstCxxRef{<:compiledDA}, CxxWrap.CxxWrapCore.CxxRef{<:compiledDA}},
    arg2::Union{AlgebraicVector{Float64}, CxxWrap.CxxWrapCore.ConstCxxRef{<:AlgebraicVector{Float64}}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{Float64}}}
)

Evaluation of arg1 with a vector of arguments, arg2.

DACE.eval — Method

eval(
    arg1::Union{compiledDA, CxxWrap.CxxWrapCore.ConstCxxRef{<:compiledDA}, CxxWrap.CxxWrapCore.CxxRef{<:compiledDA}},
    arg2::Union{CxxWrap.StdLib.StdVector{DA}, CxxWrap.CxxWrapCore.ConstCxxRef{<:CxxWrap.StdLib.StdVector{DA}}, CxxWrap.CxxWrapCore.CxxRef{<:CxxWrap.StdLib.StdVector{DA}}},
    arg3::Union{CxxWrap.StdLib.StdVector{DA}, CxxWrap.CxxWrapCore.CxxRef{<:CxxWrap.StdLib.StdVector{DA}}}
)

Evaluate the compiled polynomial, arg1, with a vector of any arithmetic type, arg2, and return vector of results, arg3.

DACE.eval — Method

eval(
    arg1::Union{compiledDA, CxxWrap.CxxWrapCore.ConstCxxRef{<:compiledDA}, CxxWrap.CxxWrapCore.CxxRef{<:compiledDA}},
    arg2::Union{CxxWrap.StdLib.StdVector{DA}, CxxWrap.CxxWrapCore.ConstCxxRef{<:CxxWrap.StdLib.StdVector{DA}}, CxxWrap.CxxWrapCore.CxxRef{<:CxxWrap.StdLib.StdVector{DA}}}
)

Evaluation of arg1 with a vector of arguments, arg2.

DACE.eval — Method

eval(
    arg1::Union{compiledDA, CxxWrap.CxxWrapCore.ConstCxxRef{<:compiledDA}, CxxWrap.CxxWrapCore.CxxRef{<:compiledDA}},
    arg2::Union{CxxWrap.StdLib.StdVector{Float64}, CxxWrap.CxxWrapCore.ConstCxxRef{<:CxxWrap.StdLib.StdVector{Float64}}, CxxWrap.CxxWrapCore.CxxRef{<:CxxWrap.StdLib.StdVector{Float64}}},
    arg3::Union{CxxWrap.StdLib.StdVector{Float64}, CxxWrap.CxxWrapCore.CxxRef{<:CxxWrap.StdLib.StdVector{Float64}}}
)

Evaluate the compiled polynomial, arg1, with a vector of any arithmetic type, arg2, and return vector of results, arg3.

DACE.eval — Method

eval(
    arg1::Union{compiledDA, CxxWrap.CxxWrapCore.ConstCxxRef{<:compiledDA}, CxxWrap.CxxWrapCore.CxxRef{<:compiledDA}},
    arg2::Union{CxxWrap.StdLib.StdVector{Float64}, CxxWrap.CxxWrapCore.ConstCxxRef{<:CxxWrap.StdLib.StdVector{Float64}}, CxxWrap.CxxWrapCore.CxxRef{<:CxxWrap.StdLib.StdVector{Float64}}}
)

Evaluation of arg1 with a vector of arguments, arg2.

DACE.evalScalar — Method

evalScalar(
    arg1::Union{AlgebraicVector{DA}, CxxWrap.CxxWrapCore.ConstCxxRef{<:AlgebraicVector{DA}}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{DA}}},
    arg2::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}}
) -> DACE.AlgebraicVectorAllocated{DA}

Evaluation of arg1 with a single arithmetic type argument, arg2.

DACE.evalScalar — Method

evalScalar(
    arg1::Union{AlgebraicVector{DA}, CxxWrap.CxxWrapCore.ConstCxxRef{<:AlgebraicVector{DA}}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{DA}}},
    arg2::Union{Float64, Int64, Ref{Float64}, Irrational}
) -> DACE.AlgebraicVectorAllocated{Float64}

Evaluation of arg1 with a single arithmetic type argument, arg2.

DACE.evalScalar — Method

evalScalar(
    arg1::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}},
    arg2::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}}
) -> DACE.DAAllocated

Evaluation of arg1 with a single arithmetic type argument, arg2.

DACE.evalScalar — Method

evalScalar(
    arg1::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}},
    arg2::Union{Float64, Int64, Ref{Float64}, Irrational}
) -> Float64

Evaluation of arg1 with a single arithmetic type argument, arg2.

DACE.evalScalar — Method

evalScalar(
    arg1::Union{compiledDA, CxxWrap.CxxWrapCore.ConstCxxRef{<:compiledDA}, CxxWrap.CxxWrapCore.CxxRef{<:compiledDA}},
    arg2::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}}
) -> CxxWrap.StdLib.StdVectorAllocated{DA}

Evaluation of arg1 with a single arithmetic type argument, arg2.

DACE.evalScalar — Method

evalScalar(
    arg1::Union{compiledDA, CxxWrap.CxxWrapCore.ConstCxxRef{<:compiledDA}, CxxWrap.CxxWrapCore.CxxRef{<:compiledDA}},
    arg2::Union{Float64, Int64, Ref{Float64}, Irrational}
) -> CxxWrap.StdLib.StdVectorAllocated{Float64}

Evaluation of arg1 with a single arithmetic type argument, arg2.

DACE.getCentralMoments — Method

getCentralMoments(
    arg1::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}},
    arg2::Integer
) -> Tuple{CxxWrap.StdLib.StdVectorAllocated{CxxWrap.StdLib.StdVector{UInt32}}, DACE.AlgebraicVectorAllocated{Float64}}

Compute the central moments up to order arg2 given the moment generating function arg1.

DACE.getCoefficient — Method

getCoefficient(
    arg1::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}},
    arg2::Union{CxxWrap.StdLib.StdVector{UInt32}, CxxWrap.CxxWrapCore.ConstCxxRef{<:CxxWrap.StdLib.StdVector{UInt32}}, CxxWrap.CxxWrapCore.CxxRef{<:CxxWrap.StdLib.StdVector{UInt32}}}
) -> Float64

Get the coefficient of arg1 with exponents arg2

DACE.getCoefficient — Method

getCoefficient(
    arg1::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}},
    arg2::Vector{UInt32}
) -> Float64

Get the coefficient of arg1 with exponents arg2

DACE.getCoefficient — Method

getCoefficient(
    arg1::Union{Monomial, CxxWrap.CxxWrapCore.ConstCxxRef{<:Monomial}, CxxWrap.CxxWrapCore.CxxRef{<:Monomial}}
) -> Float64

Get the coefficient of arg1

DACE.getCoefficient — Method

getCoefficient(
    arg1::Union{Ptr{Nothing}, CxxWrap.CxxWrapCore.ConstCxxPtr{<:DA}, CxxWrap.CxxWrapCore.CxxPtr{<:DA}},
    arg2::Union{CxxWrap.StdLib.StdVector{UInt32}, CxxWrap.CxxWrapCore.ConstCxxRef{<:CxxWrap.StdLib.StdVector{UInt32}}, CxxWrap.CxxWrapCore.CxxRef{<:CxxWrap.StdLib.StdVector{UInt32}}}
) -> Float64

Get the coefficient of arg1 with exponents arg2

DACE.getEps — Method

getEps() -> Float64

Return the cutoff value eps currently set for computations, or zero if undefined.

DACE.getEpsMac — Method

getEpsMac() -> Float64

Return the machine epsilon (pessimistic estimate), or zero if undefined.

DACE.getExponents — Method

getExponents(
    arg1::Union{Monomial, CxxWrap.CxxWrapCore.ConstCxxRef{<:Monomial}, CxxWrap.CxxWrapCore.CxxRef{<:Monomial}}
) -> CxxWrap.StdLib.StdVectorAllocated{UInt32}

Get the exponents of arg1

DACE.getMaxMonomials — Method

getMaxMonomials() -> Int64

Return the maximum number of monomials available with the order and number of variables specified, or zero if undefined.

DACE.getMaxOrder — Method

getMaxOrder() -> Int64

Return the maximum order currently set for the computation, or zero if undefined.

DACE.getMaxVariables — Method

getMaxVariables() -> Int64

Return the maximum number of variables set for the computations, or zero if undefined.

DACE.getMonomial — Method

getMonomial(
    arg1::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}},
    arg2::Integer
) -> DACE.MonomialAllocated

Get the monomial of arg1 at position arg2

DACE.getMonomials — Method

getMonomials(
    arg1::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}}
) -> CxxWrap.StdLib.StdVectorAllocated{Monomial}

Get vector of all non-zero Monomials for DA arg1

DACE.getMultiIndices — Method

getMultiIndices(
    arg1::Integer,
    arg2::Integer
) -> CxxWrap.StdLib.StdVectorAllocated{CxxWrap.StdLib.StdVector{UInt32}}

Get all multi-indices of order arg1 in arg2 variables.

DACE.getRawMoments — Method

getRawMoments(
    arg1::Union{DA, CxxWrap.CxxWrapCore.ConstCxxRef{<:DA}, CxxWrap.CxxWrapCore.CxxRef{<:DA}},
    arg2::Integer
) -> Tuple{CxxWrap.StdLib.StdVectorAllocated{CxxWrap.StdLib.StdVector{UInt32}}, DACE.AlgebraicVectorAllocated{Float64}}

Compute the raw moments up to order arg2 given the moment generating function arg1.

DACE.getTO — Method

getTO() -> Int64

Return the truncation order currently set for computations, or zero if undefined.

All terms larger than the truncation order are discarded in subsequent operations.

DACE.hessian — Method

hessian(
    arg1::Union{AlgebraicVector{DA}, CxxWrap.CxxWrapCore.ConstCxxRef{<:AlgebraicVector{DA}}, CxxWrap.CxxWrapCore.CxxRef{<:AlgebraicVector{DA}}}
) -> CxxWrap.StdLib.StdVectorAllocated{AlgebraicMatrix{DA}}

Returns a vector of Hessians where the ith entry is the Hessian of the ith element of arg1.

DACE.identity — Method

identity(arg1::Integer) -> DACE.AlgebraicVectorAllocated{DA}

Return the first arg1 DA identities.

DACE.identity — Method

identity(
    arg1::Union{CxxWrap.StdLib.StdVector{UInt32}, CxxWrap.CxxWrapCore.ConstCxxRef{<:CxxWrap.StdLib.StdVector{UInt32}}, CxxWrap.CxxWrapCore.CxxRef{<:CxxWrap.StdLib.StdVector{UInt32}}},
    arg2::Union{Bool, CxxWrap.CxxWrapCore.CxxBool}
) -> DACE.AlgebraicVectorAllocated{DA}

Return the DA identities at positions arg1, optionally sorting them if arg2 is true.

DACE.identity — Method

identity(
    arg1::Vector{UInt32},
    arg2::Union{Bool, CxxWrap.CxxWrapCore.CxxBool}
) -> DACE.AlgebraicVectorAllocated{DA}

Return the DA identities at positions arg1, optionally sorting them if arg2 is true.

DACE.init — Method

init(arg1::Integer, arg2::Integer)

Initialize the DACE control arrays and set the maximum order, arg1, and the maximum number of variables, arg2.

Note: must be called before any other DA routine can be used!

DACE.popTO — Method

popTO()

Restore the previous truncation order from the truncation order stack.

All terms larger than the truncation order are discarded in subsequent operations.

DACE.pushTO — Method

pushTO(arg1::Integer)

Set a new truncation order (arg1), saving the previous one on the truncation order stack.

All terms larger than the truncation order are discarded in subsequent operations.

DACE.setCoefficient! — Method

setCoefficient!(
    arg1::Union{DA, CxxWrap.CxxWrapCore.CxxRef{<:DA}},
    arg2::Union{CxxWrap.StdLib.StdVector{UInt32}, CxxWrap.CxxWrapCore.ConstCxxRef{<:CxxWrap.StdLib.StdVector{UInt32}}, CxxWrap.CxxWrapCore.CxxRef{<:CxxWrap.StdLib.StdVector{UInt32}}},
    arg3::Union{Float64, Int64, Irrational}
)

Set the coefficient of arg1 with exponents arg2 to arg3

DACE.setCoefficient! — Method

setCoefficient!(
    arg1::Union{DA, CxxWrap.CxxWrapCore.CxxRef{<:DA}},
    arg2::Vector{UInt32},
    arg3::Union{Float64, Int64, Irrational}
)

Set the coefficient of arg1 with exponents arg2 to arg3

DACE.setCoefficient! — Method

setCoefficient!(
    arg1::Union{Ptr{Nothing}, CxxWrap.CxxWrapCore.CxxPtr{<:DA}},
    arg2::Union{CxxWrap.StdLib.StdVector{UInt32}, CxxWrap.CxxWrapCore.ConstCxxRef{<:CxxWrap.StdLib.StdVector{UInt32}}, CxxWrap.CxxWrapCore.CxxRef{<:CxxWrap.StdLib.StdVector{UInt32}}},
    arg3::Union{Float64, Int64, Irrational}
)

Set the coefficient of arg1 with exponents arg2 to arg3

DACE.setEps — Method

setEps(arg1::Union{Float64, Int64, Irrational}) -> Float64

Set the cutoff value eps to arg1 and return the previous value, or zero if undefined.

DACE.setTO — Method

setTO(arg1::Integer) -> Int64

Set the truncation order to arg1 and return the previous value, or zero if undefined.

All terms larger than the truncation order are discarded in subsequent operations.